:sparkles: introduce stream on dynamodb table

+++ lambda/invoices/invoiceImportFunction.ts
@@ -60,24 +60,29 @@ const processRecord = async (record: S3EventRecord) => {
         const invoice = JSON.parse(object.Body!.toString('utf-8')) as InvoiceFile;
         console.log("-> invoice", invoice);
 
-       await Promise.all([
-            invoiceRepository.create({
-                pk: `#invoice_${invoice.customerName}`,
-                sk: invoice.invoiceNumber,
-                ttl: 0,
-                totalValue: invoice.totalValue,
-                productId: invoice.productId,
-                transactionId: key,
-                createdAt: Date.now(),
-                quantity: invoice.quantity,
-            }),
-            s3Client.deleteObject({
-                Key: key,
-                Bucket: record.s3.bucket.name,
-            }).promise(),
-           invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.PROCESSED),
-           invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.PROCESSED)
-        ]);
+        if (invoice.invoiceNumber.length >= 5) {
+            await Promise.all([
+                invoiceRepository.create({
+                    pk: `#invoice_${invoice.customerName}`,
+                    sk: invoice.invoiceNumber,
+                    ttl: 0,
+                    totalValue: invoice.totalValue,
+                    productId: invoice.productId,
+                    transactionId: key,
+                    createdAt: Date.now(),
+                    quantity: invoice.quantity,
+                }),
+                s3Client.deleteObject({
+                    Key: key,
+                    Bucket: record.s3.bucket.name,
+                }).promise(),
+                invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.PROCESSED),
+                invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.PROCESSED)
+            ]);
+        } else {
+            await invoiceWSService.disconnectClient(invoiceTransaction.connectionId);
+            throw new Error('Order number must be 5 or more caracter');
+        }
     } catch (error) {
         console.error(error);
     }
